Updates to Language
I forgot how huge this task is.

Along with these edits, I wanted to share a book with you. It's called "The Disappearing Spoon" and goes into the history of the Periodic Table and each of the elements contained within it: http://samkean.com/disappearing-spoon I thought maybe you'd enjoy it if you wrote a whole mission on the Periodic Table.